refactor App components into hooks
