+++ src/app/App.tsx
@@ -1,13 +1,11 @@
 // eslint-disable-next-line
 import "allotment/dist/style.css";
 
-import { useEffect, useRef } from "react";
-import { useAtom, useAtomValue, useSetAtom } from "jotai";
+import { useRef } from "react";
+import { useAtom, useAtomValue } from "jotai";
 import { Allotment, LayoutPriority } from "allotment";
 
 import styles from "./App.module.css";
-
-import { setTheme, themeMediaQuery } from "@/features/theme";
 
 import {
   currentLeftPanelAtom,
@@ -16,15 +14,17 @@ import {
   currentRightPanelAtom,
   availableLeftPanelsAtom,
 } from "@/globals/layout";
-import { themeAtom, tryUpdateThemeIfAutomaticAtom } from "@/globals/appearance";
-import { activeProjectFileAtom, useDbChangeIdCrossTabSync } from "@/globals/project";
+import { useAutomaticTheme } from "@/globals/appearance";
+import {
+  activeProjectFileAtom,
+  useDbChangeIdCrossTabSync,
+} from "@/globals/project";
 
 import AppErrorWrapperPage from "./AppErrorWrapperPage";
 import WorkspaceProvider from "./WorkspaceProvider";
 import Sidebar from "./Sidebar";
 import AppMenubar from "./AppMenubar";
 import AppStatusBar from "./AppStatusBar";
-import ProjectAutoSaver from "./ProjectAutoSaver";
 import { ToastProvider } from "@/components/Toast";
 import { TooltipProvider } from "@/components/Tooltip";
 import { DatabaseInitializer } from "./DatabaseInitializer";
@@ -44,6 +44,7 @@ import ResultTabbedPanel from "./results/ResultsTabbedPanel";
 import GraphSettingsPanel from "./graphSettings/GraphSettingsPanel";
 import ChatPanel from "./ChatPanel";
 import StartPanel from "./start/StartPanel";
+import useAutoSave from "./useAutoSave";
 
 const getDefaultResultsPanelWidth = () => {
   if (window.matchMedia && window.matchMedia("(min-width: 2000px)").matches) {
@@ -164,31 +165,10 @@ const AppContent = () => {
   );
 };
 
-const ThemeUpdater = () => {
-  const tryUpdateThemeIfAutomatic = useSetAtom(tryUpdateThemeIfAutomaticAtom);
-  const theme = useAtomValue(themeAtom);
-
-  useEffect(() => {
-    const handleChange = () => {
-      tryUpdateThemeIfAutomatic();
-    };
-
-    themeMediaQuery.addEventListener("change", handleChange);
-
-    handleChange();
-
-    return () => themeMediaQuery.removeEventListener("change", handleChange);
-  }, [tryUpdateThemeIfAutomatic]);
-
-  useEffect(() => {
-    setTheme(theme);
-  }, [theme]);
-
-  return null;
-};
-
 const AppWithHooks = () => {
   useDbChangeIdCrossTabSync();
+  useAutomaticTheme();
+  useAutoSave();
 
   return (
     <DatabaseInitializer>
@@ -201,8 +181,6 @@ const App = () => {
   return (
     <AppErrorWrapperPage>
       <AppProvider>
-        <ThemeUpdater />
-        <ProjectAutoSaver />
         <AppWithHooks />
       </AppProvider>
     </AppErrorWrapperPage>
